requirements.txt
Django==3.0
Pillow==6.2.1
django-cors-headers==3.2.0
django-filter==2.2
django-oauth-toolkit==1.2.0
djangorestframework==3.10.3
pylint==2.4.4
pycodestyle==2.5.0
python-dotenv==0.10.4
boto3==1.11.3
awscli==1.17.3
psycopg2==2.8.4
pymongo==3.10.1